+++ tests/test_pytest_cli.py
@@ -64,11 +64,30 @@ def test_simple_pytest_import_error_cli():
             """
         ''')
     temp_module = util_misc.TempModule(module_text, modname='imperr_test_mod')
-    command = sys.executable + ' -m pytest -v -s --xdoctest-verbose=3 --xdoctest ' + temp_module.dpath
+    command = sys.executable + ' -m pytest -v -s --xdoctest-verbose=3 --xdoctest-supress-import-errors --xdoctest ' + temp_module.dpath
     print(command)
+    print('--')
     info = cmd(command)
+    print('--')
+    # print('info = {}'.format(info))
+    print(info['out'])
     # We patched doctest_example so it no longer outputs this in the traceback
     assert 'util_import' not in info['out']
+    print(info['out'])
+    # Note: flaky changes the return code from 1 to 3, so test non-zero
+    assert info['ret'] != 0
+
+    # Remove the supress import error flag and now we should get the traceback
+    temp_module = util_misc.TempModule(module_text, modname='imperr_test_mod')
+    command = sys.executable + ' -m pytest -v -s --xdoctest-verbose=3 --xdoctest ' + temp_module.dpath
+    print(command)
+    print('--')
+    info = cmd(command)
+    print('--')
+    # print('info = {}'.format(info))
+    print(info['out'])
+    # We patched doctest_example so it no longer outputs this in the traceback
+    assert 'util_import' in info['out']
     print(info['out'])
     # Note: flaky changes the return code from 1 to 3, so test non-zero
     assert info['ret'] != 0